04. List Operations
The first input line will hold a list of integers. Until we receive the "End" command, we will be given operations we have to apply to the list.
The possible commands are:
· Add {number} – add the given number to the end of the list
· Insert {number} {index} – insert the number at the given index
· Remove {index} – remove the number at the given index
· Shift left {count} – first number becomes last. This has to be repeated the specified number of times
· Shift right {count} – last number becomes first. To be repeated the specified number of times
Note: the index given may be outside of the bounds of the array. In that case print: "Invalid index".