05. Bomb Numbers
Create a program that reads a sequence of numbers and a special bomb number holding a certain power. Your task is to detonate every occurrence of the special bomb number and according to its power the numbers to its left and right. The bomb power refers to how many numbers to the left and right will be removed, no matter their values. Detonations are performed from left to right and all the detonated numbers disappear. Finally, print the sum of the remaining elements in the sequence.